Moved snackbar package
