+++ app/src/main/AndroidManifest.xml
@@ -108,7 +108,7 @@
             android:label="@string/slider" />
 
         <activity
-            android:name=".activity.SnackbarActivity"
+            android:name=".components.snackbar.SnackbarActivity"
             android:label="@string/snachbar" />
 
         <activity
